fix: type

+++ src/pages/Config/Developer/useClearAllWikiData.tsx
@@ -4,16 +4,18 @@ import { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Snackbar } from 'react-native-paper';
 import { sqliteServiceService } from '../../../services/SQLiteService';
-import { IWikiWorkspace, useWorkspaceStore } from '../../../store/workspace';
+import { IWorkspace, useWorkspaceStore } from '../../../store/workspace';
 
-export const deleteWikiFile = async (wikiWorkspace: IWikiWorkspace) => {
-  await sqliteServiceService.closeDatabase(wikiWorkspace, true);
-  await fs.deleteAsync(wikiWorkspace.wikiFolderLocation);
+export const deleteWikiFile = async (wikiWorkspace: IWorkspace) => {
+  if (wikiWorkspace.type === 'wiki') {
+    await sqliteServiceService.closeDatabase(wikiWorkspace, true);
+    await fs.deleteAsync(wikiWorkspace.wikiFolderLocation);
+  }
 };
 
 export function useClearAllWikiData() {
   const { t } = useTranslation();
-  const wikis = useWorkspaceStore(state => state.workspaces);
+  const workspaces = useWorkspaceStore(state => state.workspaces);
   const removeAllWiki = useWorkspaceStore(state => state.removeAll);
 
   const [clearDataSnackBarVisible, setClearDataSnackBarVisible] = useState(false);
@@ -21,14 +23,14 @@ export function useClearAllWikiData() {
   const clearAllWikiData = useCallback(async () => {
     try {
       // eslint-disable-next-line unicorn/no-array-callback-reference
-      await Promise.all(wikis.map(deleteWikiFile));
+      await Promise.all(workspaces.map(deleteWikiFile));
       removeAllWiki();
       setClearDataSnackBarVisible(true);
     } catch (error) {
       setClearDataSnackBarVisible(true);
       setClearDataSnackBarErrorMessage((error as Error).message);
     }
-  }, [removeAllWiki, wikis]);
+  }, [removeAllWiki, workspaces]);
 
   const ClearAllWikiDataResultSnackBar = (
     <Snackbar
